+++ src/main/java/it/gov/pagopa/wispconverter/config/client/APIConfigCacheClientConfig.java
@@ -7,6 +7,7 @@ import it.gov.pagopa.wispconverter.util.client.apiconfigcache.ApiConfigCacheClie
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.client.BufferingClientHttpRequestFactory;
@@ -36,42 +37,28 @@ public class APIConfigCacheClientConfig {
     @Value("${client.cache.api-key}")
     private String apiKey;
 
-    @Value("${log.client.cache.request.include-headers}")
-    private boolean clientRequestIncludeHeaders;
-    @Value("${log.client.cache.request.include-payload}")
-    private boolean clientRequestIncludePayload;
-    @Value("${log.client.cache.request.max-payload-length}")
-    private int clientRequestMaxLength;
-    @Value("${log.client.cache.response.include-headers}")
-    private boolean clientResponseIncludeHeaders;
-    @Value("${log.client.cache.response.include-payload}")
-    private boolean clientResponseIncludePayload;
-    @Value("${log.client.cache.response.max-payload-length}")
-    private int clientResponseMaxLength;
-
-    @Value("${log.client.cache.mask.header.name}")
-    private String maskHeaderName;
-
-    @Value("${log.client.cache.request.pretty}")
-    private boolean clientRequestPretty;
-
-    @Value("${log.client.cache.response.pretty}")
-    private boolean clientResponsePretty;
+    @Bean
+    @ConfigurationProperties(prefix = "log.client.cache")
+    public ClientLoggingProperties cacheClientLoggingProperties() {
+        return new ClientLoggingProperties();
+    }
 
 
     @Bean
     public it.gov.pagopa.gen.wispconverter.client.cache.invoker.ApiClient configCacheClient() {
-        ApiConfigCacheClientLoggingInterceptor clientLogging = new ApiConfigCacheClientLoggingInterceptor(reService);
-        clientLogging.setRequestIncludeHeaders(clientRequestIncludeHeaders);
-        clientLogging.setRequestIncludePayload(clientRequestIncludePayload);
-        clientLogging.setRequestMaxPayloadLength(clientRequestMaxLength);
-        clientLogging.setRequestHeaderPredicate(p -> !p.equals(maskHeaderName));
-        clientLogging.setRequestPretty(clientRequestPretty);
+        ClientLoggingProperties clientLoggingProperties = cacheClientLoggingProperties();
 
-        clientLogging.setResponseIncludeHeaders(clientResponseIncludeHeaders);
-        clientLogging.setResponseIncludePayload(clientResponseIncludePayload);
-        clientLogging.setResponseMaxPayloadLength(clientResponseMaxLength);
-        clientLogging.setResponsePretty(clientResponsePretty);
+        ApiConfigCacheClientLoggingInterceptor clientLogging = new ApiConfigCacheClientLoggingInterceptor(reService);
+        clientLogging.setRequestIncludeHeaders(clientLoggingProperties.getRequest().isIncludeHeaders());
+        clientLogging.setRequestIncludePayload(clientLoggingProperties.getRequest().isIncludePayload());
+        clientLogging.setRequestMaxPayloadLength(clientLoggingProperties.getRequest().getMaxPayloadLength());
+        clientLogging.setRequestHeaderPredicate(p -> !p.equals(clientLoggingProperties.getRequest().getMaskHeaderName()));
+        clientLogging.setRequestPretty(clientLoggingProperties.getRequest().isPretty());
+
+        clientLogging.setResponseIncludeHeaders(clientLoggingProperties.getResponse().isIncludeHeaders());
+        clientLogging.setResponseIncludePayload(clientLoggingProperties.getResponse().isIncludePayload());
+        clientLogging.setResponseMaxPayloadLength(clientLoggingProperties.getResponse().getMaxPayloadLength());
+        clientLogging.setResponsePretty(clientLoggingProperties.getResponse().isPretty());
 
         RestTemplate restTemplate = restTemplate();
