+++ src/main/java/it/gov/pagopa/wispconverter/config/client/DecouplerCachingClientConfig.java
@@ -7,6 +7,7 @@ import it.gov.pagopa.wispconverter.util.client.decouplercaching.DecouplerCaching
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.client.BufferingClientHttpRequestFactory;
@@ -37,41 +38,28 @@ public class DecouplerCachingClientConfig {
     @Value("${client.decoupler-caching.api-key}")
     private String apiKey;
 
-    @Value("${log.client.decoupler-caching.request.include-headers}")
-    private boolean clientRequestIncludeHeaders;
-    @Value("${log.client.decoupler-caching.request.include-payload}")
-    private boolean clientRequestIncludePayload;
-    @Value("${log.client.decoupler-caching.request.max-payload-length}")
-    private int clientRequestMaxLength;
-    @Value("${log.client.decoupler-caching.response.include-headers}")
-    private boolean clientResponseIncludeHeaders;
-    @Value("${log.client.decoupler-caching.response.include-payload}")
-    private boolean clientResponseIncludePayload;
-    @Value("${log.client.decoupler-caching.response.max-payload-length}")
-    private int clientResponseMaxLength;
 
-    @Value("${log.client.decoupler-caching.mask.header.name}")
-    private String maskHeaderName;
-
-    @Value("${log.client.decoupler-caching.request.pretty}")
-    private boolean clientRequestPretty;
-
-    @Value("${log.client.decoupler-caching.response.pretty}")
-    private boolean clientResponsePretty;
+    @Bean
+    @ConfigurationProperties(prefix = "log.client.decoupler-caching")
+    public ClientLoggingProperties decouplerCachingClientLoggingProperties() {
+        return new ClientLoggingProperties();
+    }
 
     @Bean
     public it.gov.pagopa.gen.wispconverter.client.decouplercaching.invoker.ApiClient decouplerCachingClient() {
-        DecouplerCachingClientLoggingInterceptor clientLogging = new DecouplerCachingClientLoggingInterceptor(reService);
-        clientLogging.setRequestIncludeHeaders(clientRequestIncludeHeaders);
-        clientLogging.setRequestIncludePayload(clientRequestIncludePayload);
-        clientLogging.setRequestMaxPayloadLength(clientRequestMaxLength);
-        clientLogging.setRequestHeaderPredicate(p -> !p.equals(maskHeaderName));
-        clientLogging.setRequestPretty(clientRequestPretty);
+        ClientLoggingProperties clientLoggingProperties = decouplerCachingClientLoggingProperties();
 
-        clientLogging.setResponseIncludeHeaders(clientResponseIncludeHeaders);
-        clientLogging.setResponseIncludePayload(clientResponseIncludePayload);
-        clientLogging.setResponseMaxPayloadLength(clientResponseMaxLength);
-        clientLogging.setResponsePretty(clientResponsePretty);
+        DecouplerCachingClientLoggingInterceptor clientLogging = new DecouplerCachingClientLoggingInterceptor(reService);
+        clientLogging.setRequestIncludeHeaders(clientLoggingProperties.getRequest().isIncludeHeaders());
+        clientLogging.setRequestIncludePayload(clientLoggingProperties.getRequest().isIncludePayload());
+        clientLogging.setRequestMaxPayloadLength(clientLoggingProperties.getRequest().getMaxPayloadLength());
+        clientLogging.setRequestHeaderPredicate(p -> !p.equals(clientLoggingProperties.getRequest().getMaskHeaderName()));
+        clientLogging.setRequestPretty(clientLoggingProperties.getRequest().isPretty());
+
+        clientLogging.setResponseIncludeHeaders(clientLoggingProperties.getResponse().isIncludeHeaders());
+        clientLogging.setResponseIncludePayload(clientLoggingProperties.getResponse().isIncludePayload());
+        clientLogging.setResponseMaxPayloadLength(clientLoggingProperties.getResponse().getMaxPayloadLength());
+        clientLogging.setResponsePretty(clientLoggingProperties.getResponse().isPretty());
 
         RestTemplate restTemplate = restTemplate();
