+++ src/main/java/it/gov/pagopa/wispconverter/config/client/IuvGeneratorClientConfig.java
@@ -7,6 +7,7 @@ import it.gov.pagopa.wispconverter.util.client.iuvgenerator.IuvGeneratorClientRe
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.client.BufferingClientHttpRequestFactory;
@@ -37,42 +38,29 @@ public class IuvGeneratorClientConfig {
     @Value("${client.iuvgenerator.api-key}")
     private String apiKey;
 
-    @Value("${log.client.iuvgenerator.request.include-headers}")
-    private boolean clientRequestIncludeHeaders;
-    @Value("${log.client.iuvgenerator.request.include-payload}")
-    private boolean clientRequestIncludePayload;
-    @Value("${log.client.iuvgenerator.request.max-payload-length}")
-    private int clientRequestMaxLength;
-    @Value("${log.client.iuvgenerator.response.include-headers}")
-    private boolean clientResponseIncludeHeaders;
-    @Value("${log.client.iuvgenerator.response.include-payload}")
-    private boolean clientResponseIncludePayload;
-    @Value("${log.client.iuvgenerator.response.max-payload-length}")
-    private int clientResponseMaxLength;
 
-    @Value("${log.client.iuvgenerator.mask.header.name}")
-    private String maskHeaderName;
-
-    @Value("${log.client.iuvgenerator.request.pretty}")
-    private boolean clientRequestPretty;
-
-    @Value("${log.client.iuvgenerator.response.pretty}")
-    private boolean clientResponsePretty;
+    @Bean
+    @ConfigurationProperties(prefix = "log.client.iuvgenerator")
+    public ClientLoggingProperties iuvGeneratorClientLoggingProperties() {
+        return new ClientLoggingProperties();
+    }
 
 
     @Bean
     public it.gov.pagopa.gen.wispconverter.client.iuvgenerator.invoker.ApiClient iuvGeneratorClient() {
-        IuvGeneratorClientLoggingInterceptor clientLogging = new IuvGeneratorClientLoggingInterceptor(reService);
-        clientLogging.setRequestIncludeHeaders(clientRequestIncludeHeaders);
-        clientLogging.setRequestIncludePayload(clientRequestIncludePayload);
-        clientLogging.setRequestMaxPayloadLength(clientRequestMaxLength);
-        clientLogging.setRequestHeaderPredicate(p -> !p.equals(maskHeaderName));
-        clientLogging.setRequestPretty(clientRequestPretty);
+        ClientLoggingProperties clientLoggingProperties = iuvGeneratorClientLoggingProperties();
 
-        clientLogging.setResponseIncludeHeaders(clientResponseIncludeHeaders);
-        clientLogging.setResponseIncludePayload(clientResponseIncludePayload);
-        clientLogging.setResponseMaxPayloadLength(clientResponseMaxLength);
-        clientLogging.setResponsePretty(clientResponsePretty);
+        IuvGeneratorClientLoggingInterceptor clientLogging = new IuvGeneratorClientLoggingInterceptor(reService);
+        clientLogging.setRequestIncludeHeaders(clientLoggingProperties.getRequest().isIncludeHeaders());
+        clientLogging.setRequestIncludePayload(clientLoggingProperties.getRequest().isIncludePayload());
+        clientLogging.setRequestMaxPayloadLength(clientLoggingProperties.getRequest().getMaxPayloadLength());
+        clientLogging.setRequestHeaderPredicate(p -> !p.equals(clientLoggingProperties.getRequest().getMaskHeaderName()));
+        clientLogging.setRequestPretty(clientLoggingProperties.getRequest().isPretty());
+
+        clientLogging.setResponseIncludeHeaders(clientLoggingProperties.getResponse().isIncludeHeaders());
+        clientLogging.setResponseIncludePayload(clientLoggingProperties.getResponse().isIncludePayload());
+        clientLogging.setResponseMaxPayloadLength(clientLoggingProperties.getResponse().getMaxPayloadLength());
+        clientLogging.setResponsePretty(clientLoggingProperties.getResponse().isPretty());
 
         RestTemplate restTemplate = restTemplate();
