+++ src/main/java/it/gov/pagopa/wispconverter/service/model/re/CallTypeEnum.java
@@ -1,0 +1,6 @@
+package it.gov.pagopa.wispconverter.service.model.re;
+
+public enum CallTypeEnum {
+    SERVER,
+    CLIENT;
+}
