+++ src/main/resources/application.properties
@@ -113,7 +113,7 @@ log.client.checkout.request.max-payload-length=10000
 log.client.checkout.response.include-headers=true
 log.client.checkout.response.include-payload=true
 log.client.checkout.response.max-payload-length=10000
-log.client.checkout.mask.header.name=${CLIENT_CHECKOUT_API_KEY_NAME:Ocp-Apim-Subscription-Key}
+log.client.checkout.request.mask-header-name=${CLIENT_CHECKOUT_API_KEY_NAME:Ocp-Apim-Subscription-Key}
 log.client.checkout.request.pretty=true
 log.client.checkout.response.pretty=true
 
@@ -125,7 +125,7 @@ log.client.decoupler-caching.request.max-payload-length=10000
 log.client.decoupler-caching.response.include-headers=true
 log.client.decoupler-caching.response.include-payload=true
 log.client.decoupler-caching.response.max-payload-length=10000
-log.client.decoupler-caching.mask.header.name=${CLIENT_DECOUPLERCACHING_API_KEY_NAME:Ocp-Apim-Subscription-Key}
+log.client.decoupler-caching.request.mask-header-name=${CLIENT_DECOUPLERCACHING_API_KEY_NAME:Ocp-Apim-Subscription-Key}
 log.client.decoupler-caching.request.pretty=true
 log.client.decoupler-caching.response.pretty=true
 
@@ -137,7 +137,7 @@ log.client.iuvgenerator.request.max-payload-length=10000
 log.client.iuvgenerator.response.include-headers=true
 log.client.iuvgenerator.response.include-payload=true
 log.client.iuvgenerator.response.max-payload-length=10000
-log.client.iuvgenerator.mask.header.name=${CLIENT_IUVGENERATOR_API_KEY_NAME:Ocp-Apim-Subscription-Key}
+log.client.iuvgenerator.request.mask-header-name=${CLIENT_IUVGENERATOR_API_KEY_NAME:Ocp-Apim-Subscription-Key}
 log.client.iuvgenerator.request.pretty=true
 log.client.iuvgenerator.response.pretty=true
 
@@ -149,7 +149,7 @@ log.client.gpd.request.max-payload-length=10000
 log.client.gpd.response.include-headers=true
 log.client.gpd.response.include-payload=true
 log.client.gpd.response.max-payload-length=10000
-log.client.gpd.mask.header.name=${CLIENT_GPD_API_KEY_NAME:Ocp-Apim-Subscription-Key}
+log.client.gpd.request.mask-header-name=${CLIENT_GPD_API_KEY_NAME:Ocp-Apim-Subscription-Key}
 log.client.gpd.request.pretty=true
 log.client.gpd.response.pretty=true
 
@@ -161,6 +161,6 @@ log.client.cache.request.max-payload-length=10000
 log.client.cache.response.include-headers=true
 log.client.cache.response.include-payload=true
 log.client.cache.response.max-payload-length=10000
-log.client.cache.mask.header.name=${CLIENT_CONFIG_API_KEY_NAME:Ocp-Apim-Subscription-Key}
+log.client.cache.request.mask-header-name=${CLIENT_CONFIG_API_KEY_NAME:Ocp-Apim-Subscription-Key}
 log.client.cache.request.pretty=true
 log.client.cache.response.pretty=true
